pyproject.toml
[project]
name = "llamaindex-chat"
version = "0.1.0"
description = ""
authors = [
    { name = "None" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "openai (>=1.61.1,<2.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pandas-stubs (>=2.2.3.241126,<3.0.0.0)",
    "pypdf (>=5.3.0,<6.0.0)",
    "llama-index (>=0.12.16,<0.13.0)",
    "langchain-openai (>=0.3.4,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false


[tool.mypy]
python_version = "3.13"
strict = true
namespace_packages = true
exclude = "(^|/|\\\\)venv($|/|\\\\)"
[[tool.mypy.overrides]]
module = "llama_index.*"
ignore_missing_imports = true
